requirements.txt
fastapi==0.100.0
Flask==2.3.0
flask-cors==3.0.10
flask-sockets==0.2.1
gevent-websocket==0.10.1
openai==0.27.4
python-dotenv==1.0.0
PyJWT==2.3.0
uvicorn==0.22.0
websockets==11.0.3
starlette~=0.27.0
aiosqlite==0.20.0
bcrypt==4.2.1